refactor(validator): record structure
